Create diabetes_Group_5.R
The objective of the dataset is to diagnostically predict
whether or not a patient has diabetes, based on certain
diagnostic measurements included in the dataset. The
dataset gives us details about the various differences
in the circumstanvces faced by the wide variety of
subjects .it focuses on a few basic important
parameters like pregnancies,glucose and blood
pressure etc. The main objective of data set is to
classify that wether a perticular person has diabetes or
not with help of several independaent variables given
in data set.

by- Kriti Banka , Ayush Agarwal and Aaryan Bhardwaj